MAJ 1er résultats + ajout fichier de fonctions
ajout d'un fichier regroupant les fonctions utiles :
-read_header : reads header of a wav file to get info such as duration, samplerate etc...
-extract_datetime : extract datetime from filename such as Apocado / Cetiroise or custom ones
- sorting_annot_boxes : From an Aplose results csv, returns a DataFrame without the Aplose box annotations (weak annotations)
-t_rounder : Rounds to nearest 10-minute interval
- from_str2ts : from APLOSE date string to a timestamp
- from_str2dt : from APLOSE date string to a datetime
